refactor activate/deactivate
